+++ java/src/main/java/com/stackmob/example/UpdateRelationships.java
@@ -213,17 +213,23 @@ public class UpdateRelationships implements CustomCodeMethod {
 						// no need to remove if any of the types is already block or delete
 						if (type >= 3L && typeUser.getValue().longValue() < 3L && typeOther.getValue().longValue() < 3L) {
 							if (relObject.getValue().containsKey("events_by_owner")) {
-								SMList<SMString> events = (SMList<SMString>)relObject.getValue().get("events_by_owner");
-								dataService.removeRelatedObjects("relationship", relId, "events_by_owner", events, true);
+								SMList<SMString> eventsValue = (SMList<SMString>)relObject.getValue().get("events_by_owner");
+								List<SMString> events = eventsValue.getValue();
+								if (events.size() > 0) {
+									dataService.removeRelatedObjects("relationship", relId, "events_by_owner", events, true);
+								}
 								if (userRole.equals("receiver")) {
-									removedEventIds.addAll(events.getValue());
+									removedEventIds.addAll(events);
 								}
 							}
 							if (relObject.getValue().containsKey("events_by_receiver")) {
-								SMList<SMString> events = (SMList<SMString>)relObject.getValue().get("events_by_receiver");
-								dataService.removeRelatedObjects("relationship", relId, "events_by_receiver", events, true);
+								SMList<SMString> eventsValue = (SMList<SMString>)relObject.getValue().get("events_by_receiver");
+								List<SMString> events = eventsValue.getValue();
+								if (events.size() > 0) {
+									dataService.removeRelatedObjects("relationship", relId, "events_by_receiver", events, true);
+								}
 								if (userRole.equals("owner")) {
-									removedEventIds.addAll(events.getValue());
+									removedEventIds.addAll(events);
 								}
 							}
 						}
@@ -254,8 +260,10 @@ public class UpdateRelationships implements CustomCodeMethod {
 									}
 								}
 								// remove groups from relationship's groups by user
-								dataService.removeRelatedObjects("relationship", relId, groupKey, groupIdList, false);
-								groupChange = true;
+								if (groupIdList.size() > 0) {
+									dataService.removeRelatedObjects("relationship", relId, groupKey, groupIdList, false);
+									groupChange = true;
+								}
 							}
 						}
 						// update type by user
